BUGFIX: next level cache should be accessed on every miss.
Data from the next level cache should be read in on every miss,
regardless of the read/write nature of the access. This is because you
are evicting an *entire cache entry*, not just the address that was
orignally being requested. Yes, on a write access, you have the valid
data *for that address*, so there is no need to read from the next level
*for that address*. But you don't have a clue as to what should go into
your other data blocks for that cache entry.

With this fix, the code passes all the provided validation data for the
parts that are implemented (i.e., all read/write/miss counts are
correct, and AAT is correct).